+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `--dry-run` mode to print out what it normally would run otherwise
+
 ## [0.0.1] - 2020-08-22
 
 ### Added